requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg-binary
psycopg
python-jose[cryptography]
requests
